# Optimize claim-token-reward execution costs