--- a/src/citycoin-client.ts
+++ b/src/citycoin-client.ts
@@ -102,18 +102,18 @@ export class CityCoinClient {
     return this.callReadOnlyFn("get-pox-lite-info");
   }
 
-  getBlockCommitTotal(miners: MinersList): Result {
+  getBlockCommitTotal(stacksBlockHeight: number): Result {
     return this.callReadOnlyFn(
       "get-block-commit-total",
       [
-        miners.convert()
+        types.uint(stacksBlockHeight)
       ]
     );
   }
 
   getBlockWinner(stacksBlockHeight: number, randomSampleUint: number, miners: MinersList): Result {
     return this.callReadOnlyFn(
-      "get-block-winner-imp",
+      "get-block-winner",
       [
         types.uint(stacksBlockHeight),
         types.uint(randomSampleUint),