… contracts/test_addons/citycoin.clar