This PR contains draft changes that optimize execution costs of `claim-token-reward` function.

I added new map `block-commit` (not a great name, but good enough for PoC) that stores sum of all commitments placed by miners at specific block. Values in it are updated when miners calls `mine-tokens` - by doing so we always have sum of all commitment ready to be used and we don't have to use expensive `fold` to calculate it when we need it.
This solution slightly increases the execution costs of  `mine-tokens` but reduces the cost of `claim-token-reward` by ~30%.

Second part of this PoC are new implementations of `get-block-winner` and `get-block-winner-closure` named accordingly `get-block-winner-imp` and `get-block-winner-closure`.
Original design use static, predefined list of unsigned integers `REWARD-CYCLE-INDEXES ` as input list in `fold` function and at each "iteration" passes list with miners and their commitments. It makes it inefficient when we have only few miners, as every single time `fold` will execute same number of "iterations" (32 with current miners limit).

I decided to try a bit more dynamic approach, where we pass list with miners as an input list to `fold` function, and reduce data used in each iteration to bare minimum.


Execution costs were acquired as follows:

```sh
# test with only one miner (this one was used to gather BEFORE data)
$ npm run clarinet:console
> (contract-call? .citycoin register-miner)
> ::advance_chain_tip 200
> (contract-call? .citycoin mine-tokens u200)
> ::advance_chain_tip 200
> ::get_costs (contract-call? .citycoin claim-token-reward u200)

# test with 32 miners and our miner joined as the last one
$ npm run clarinet:console
> (contract-call? .citycoin register-miner)
> ::advance_chain_tip 200
> (contract-call? .citycoin setup-31-miners u2)
> (contract-call? .citycoin mine-tokens u200)
> ::advance_chain_tip 200
> ::get_costs (contract-call? .citycoin claim-token-reward u200)

# test with 32 miners and our miner joined as the first one
$ npm run clarinet:console
> (contract-call? .citycoin register-miner)
> ::advance_chain_tip 200
> (contract-call? .citycoin mine-tokens u200)
> (contract-call? .citycoin setup-31-miners u2)
> ::advance_chain_tip 200
> ::get_costs (contract-call? .citycoin claim-token-reward u200)
``` 

**BEFORE**
```
+----------------------+------------+------------+
|                      | Consumed   | Limit      |
+----------------------+------------+------------+
| Runtime              | 2747183000 | 5000000000 |
+----------------------+------------+------------+
| Read count           | 14         | 7750       |
+----------------------+------------+------------+
| Read length (bytes)  | 45455      | 100000000  |
+----------------------+------------+------------+
| Write count          | 3          | 7750       |
+----------------------+------------+------------+
| Write length (bytes) | 6743       | 15000000   |
+----------------------+------------+------------+
```

```AFTER
+----------------------+-----------+------------+
|                      | Consumed  | Limit      |
+----------------------+-----------+------------+
| Runtime              | 145607000 - 210486000 | 5000000000 |
+----------------------+-----------+------------+
| Read count           | 15        | 7750       |
+----------------------+-----------+------------+
| Read length (bytes)  | 47451     | 100000000  |
+----------------------+-----------+------------+
| Write count          | 3         | 7750       |
+----------------------+-----------+------------+
| Write length (bytes) | 6743      | 15000000   |
+----------------------+-----------+------------+
```

Together both changes reduced execution costs somewhere between 13x and 18.8x (depending on how many miners participated in mining at specific block).

In final solution most likely `block-commit` should be merged with `miners` map, to save some space.

Before I fix native tests (initial test suite created by @jcnelson) and  bring this change to its final shape, I'd like to get some feedback from you guys.
What do you think of this approach?
Do you find any potential issues we might face in the future with it?