@whoabuddy 2021-05-25
I think this is great, we need those costs to be as low as possible to look at either increasing the number of miners or how to implement mining across multiple blocks. 👍 from me!